sender/meteofrance2openhasp/requirements.txt
meteofrance-api>=1.0.2
paho-mqtt>=2.1.0
PyYAML>=6.0.2
jinja2>=3.1.5
